+++ CHANGELOG.md
@@ -1,3 +1,14 @@
 # Image Size Changelog
 
-## [Initial Version] - {PR_MERGE_DATE}
+## [0.2.0] - 2025-03-21
+
+### Added
+
+- New "Image Resize Calculator" command
+- Calculate new dimensions while maintaining aspect ratio
+- Enter either width or height to calculate the other dimension
+- Automatically copy calculated dimension to clipboard
+
+## [0.1.0] - Initial Release
+
+- Display image dimensions and copy width or height to clipboard